AddDialySalesの構成変更
